+++ src/test/java/btrplace/plan/action/BootVMTest.java
@@ -43,7 +43,7 @@ public class BootVMTest {
         Assert.assertEquals(n, a.getDestinationNode());
         Assert.assertEquals(3, a.getStart());
         Assert.assertEquals(5, a.getEnd());
-        Assert.assertNotNull(a);
+        Assert.assertFalse(a.toString().contains("null"));
     }
 
     @Test(dependsOnMethods = {"testInstantiate"})