+++ src/test/java/btrplace/plan/action/InstantiateVMTest.java
@@ -41,7 +41,7 @@ public class InstantiateVMTest {
         Assert.assertEquals(vm, a.getVM());
         Assert.assertEquals(3, a.getStart());
         Assert.assertEquals(5, a.getEnd());
-        Assert.assertNotNull(a);
+        Assert.assertFalse(a.toString().contains("null"));
     }
 
     @Test(dependsOnMethods = {"testInstantiate"})