+++ src/test/java/btrplace/plan/action/ShutdownVMTest.java
@@ -43,7 +43,8 @@ public class ShutdownVMTest {
         Assert.assertEquals(n, a.getNode());
         Assert.assertEquals(3, a.getStart());
         Assert.assertEquals(5, a.getEnd());
-        Assert.assertNotNull(a);
+        Assert.assertFalse(a.toString().contains("null"));
+
     }
 
     @Test(dependsOnMethods = {"testInstantiate"})